+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Changes
 ```diff
 + Encryption queue (backpressure especially for Argon2 and BCrypt hashing methods which can lead to OOM killer)
++ Argon2 configuration settings
 ```
 
 ### Fixes